MOB-1047: Add edit totp

+++ domain/dto/src/main/java/com/passbolt/mobile/android/dto/request/CreateResourceDto.kt
@@ -54,6 +54,6 @@ data class TotpSecretsDto(
         val algorithm: String,
         val key: String,
         val digits: Int,
-        val period: Int
+        val period: Long
     )
 }
