+++ features/otp/src/main/java/com/passbolt/mobile/android/feature/otp/createotpmanuallyexpertsettings/CreateOtpAdvancedSettingsPresenter.kt
@@ -65,7 +65,7 @@ class CreateOtpAdvancedSettingsPresenter : CreateOtpAdvancedSettingsContract.Pre
             onValid {
                 view?.applyChangesAndGoBack(
                     OtpAdvancedSettingsModel(
-                        period.toInt(), algorithm, digits.toInt()
+                        period.toLong(), algorithm, digits.toInt()
                     )
                 )
             }
